requirements.txt
flask

flask-cors

gevent

pytorch-pretrained-bert>=0.6
